+++ server/routes/termdb.cluster.ts
@@ -1,24 +1,16 @@
 import path from 'path'
 import run_R from '#src/run_R.js'
 import type {
-	TermdbClusterRequestGeneExpression,
 	TermdbClusterRequest,
 	TermdbClusterResponse,
 	Clustering,
 	ValidResponse,
 	SingletermResponse,
-	GeneExpressionQuery,
-	GeneExpressionQueryNative,
-	GeneExpressionQueryGdc,
 	RouteApi
 } from '#types'
 import { termdbClusterPayload } from '#types/checkers'
-import * as utils from '#src/utils.js'
 import serverconfig from '#src/serverconfig.js'
-import { gdc_validate_query_geneExpression } from '#src/mds3.gdc.js'
-import { mayLimitSamples } from '#src/mds3.filter.js'
 import { clusterMethodLst, distanceMethodLst } from '#shared/clustering.js'
-import { getResult as getResultGene } from '#src/gene.js'
 import { TermTypes, NUMERIC_DICTIONARY_TERM } from '#shared/terms.js'
 import { getData } from '#src/termdb.matrix.js'
 import { termType2label } from '#shared/terms.js'
@@ -236,114 +228,4 @@ function getZscore(l: number[]) {
 		return l
 	}
 	return l.map(v => (v - mean) / sd)
-}
-
-export async function validate_query_geneExpression(ds: any, genome: any) {
-	const q: GeneExpressionQuery = ds.queries.geneExpression
-	if (!q) return
-	q.geneExpression2bins = {} //this dict is used to store the default bin config for each gene searched, so it doesn't have to be recalculated each time
-
-	if (q.src == 'gdcapi') {
-		gdc_validate_query_geneExpression(ds as GeneExpressionQueryGdc, genome)
-		// q.get() added
-		return
-	}
-	if (q.src == 'native') {
-		await validateNative(q as GeneExpressionQueryNative, ds, genome)
-		return
-	}
-	throw 'unknown queries.geneExpression.src'
-}
-
-async function validateNative(q: GeneExpressionQueryNative, ds: any, genome: any) {
-	if (!q.file.startsWith(serverconfig.tpmasterdir)) q.file = path.join(serverconfig.tpmasterdir, q.file)
-	if (!q.samples) q.samples = []
-	await utils.validate_tabixfile(q.file)
-	q.nochr = await utils.tabix_is_nochr(q.file, null, genome)
-	q.samples = [] as number[]
-
-	{
-		// is a gene-by-sample matrix file
-		const lines = await utils.get_header_tabix(q.file)
-		if (!lines[0]) throw 'header line missing from ' + q.file
-		const l = lines[0].split('\t')
-		if (l.slice(0, 4).join('\t') != '#chr\tstart\tstop\tgene') throw 'header line has wrong content for columns 1-4'
-		for (let i = 4; i < l.length; i++) {
-			const id = ds.cohort.termdb.q.sampleName2id(l[i])
-			if (id == undefined) throw 'queries.geneExpression: unknown sample from header: ' + l[i]
-			q.samples.push(id)
-		}
-	}
-
-	q.get = async (param: TermdbClusterRequestGeneExpression) => {
-		const limitSamples = await mayLimitSamples(param, q.samples, ds)
-		if (limitSamples?.size == 0) {
-			// got 0 sample after filtering, must still return expected structure with no data
-			return { term2sample2value: new Map(), byTermId: {}, bySampleId: {} }
-		}
-
-		// has at least 1 sample passing filter and with exp data
-		// TODO what if there's just 1 sample not enough for clustering?
-		const bySampleId = {}
-		const samples = q.samples || []
-		if (limitSamples) {
-			for (const sid of limitSamples) {
-				bySampleId[sid] = { label: ds.cohort.termdb.q.id2sampleName(sid) }
-			}
-		} else {
-			// use all samples with exp data
-			for (const sid of samples) {
-				bySampleId[sid] = { label: ds.cohort.termdb.q.id2sampleName(sid) }
-			}
-		}
-
-		// only valid genes with data are added. invalid genes or genes missing from data file is not added. backend returned genes is allowed to be fewer than supplied by client
-		const term2sample2value = new Map() // k: gene symbol, v: { sampleId : value }
-
-		for (const geneTerm of param.terms) {
-			if (!geneTerm.gene) continue
-			if (!geneTerm.chr || !Number.isInteger(geneTerm.start) || !Number.isInteger(geneTerm.stop)) {
-				// need to supply chr/start/stop to query
-				// legacy fpkm files
-				// will not be necessary once these files are retired
-				const re = getResultGene(genome, { input: geneTerm.gene, deep: 1 })
-				if (!re.gmlst || re.gmlst.length == 0) {
-					console.warn('unknown gene:' + geneTerm.gene) // TODO unknown genes should be notified to client
-					continue
-				}
-				const i = re.gmlst.find(i => i.isdefault) || re.gmlst[0]
-				geneTerm.start = i.start
-				geneTerm.stop = i.stop
-				geneTerm.chr = i.chr
-			}
-
-			const s2v = {}
-			if (!geneTerm.chr || !Number.isInteger(geneTerm.start) || !Number.isInteger(geneTerm.stop))
-				throw 'missing chr/start/stop'
-			await utils.get_lines_bigfile({
-				args: [
-					q.file,
-					(q.nochr ? geneTerm.chr.replace('chr', '') : geneTerm.chr) + ':' + geneTerm.start + '-' + geneTerm.stop
-				],
-				callback: line => {
-					const l = line.split('\t')
-					// case-insensitive match! FIXME if g.gene is alias won't work
-					if (l[3].toLowerCase() != geneTerm.gene.toLowerCase()) return
-					for (let i = 4; i < l.length; i++) {
-						const sampleId = samples[i - 4]
-						if (limitSamples && !limitSamples.has(sampleId)) continue // doing filtering and sample of current column is not used
-						if (!l[i]) continue // blank string
-						const v = Number(l[i])
-						if (Number.isNaN(v)) throw 'exp value not number'
-						s2v[sampleId] = v
-					}
-				}
-			})
-			if (Object.keys(s2v).length) term2sample2value.set(geneTerm.gene, s2v) // only add gene if has data
-		}
-		// pass blank byTermId to match with expected output structure
-		const byTermId = {}
-		if (term2sample2value.size == 0) throw 'no data available for the input ' + param.terms?.map(g => g.gene).join(', ')
-		return { term2sample2value, byTermId, bySampleId }
-	}
 }
