+++ server/src/mds3.init.js
@@ -36,7 +36,7 @@ import { validate_correlationVolcano } from '#routes/correlationVolcano.ts'
 import { validate_query_singleCell } from '#routes/termdb.singlecellSamples.ts'
 import { validate_query_TopVariablyExpressedGenes } from '#routes/termdb.topVariablyExpressedGenes.ts'
 import { validate_query_singleSampleMutation } from '#routes/termdb.singleSampleMutation.ts'
-import { validate_query_geneExpression } from '#routes/termdb.cluster.ts'
+import { validate_query_geneExpression } from '#routes/termdb.DE.ts'
 import { mayLimitSamples, tid2value2filter } from './mds3.filter.js'
 import { getResult } from '#src/gene.js'
 import { validate_query_getTopTermsByType } from '#routes/termdb.topTermsByType.ts'
